create and use ConcretePublishableFactory
