+++ publish/tests/factories.py
@@ -1,6 +1,9 @@
 import factory
 
 from django.contrib.auth import get_user_model
+from django.db import connection
+
+from ..models import Publishable
 
 
 class UserFactory(factory.django.DjangoModelFactory):
@@ -9,3 +12,21 @@ class UserFactory(factory.django.DjangoModelFactory):
 
     username = factory.Sequence(lambda n: 'User %d' % n)
     password = factory.Sequence(lambda n: 'password%d' % n)
+
+
+def setupModels(*models):
+    with connection.schema_editor() as schema_editor:
+        for model in models:
+            schema_editor.create_model(model)
+
+
+class ConcretePublishableModel(Publishable):
+    """A concrete model for testing the Publishable model; only available in the test suite."""
+    class Meta:
+        app_label = 'publish'
+
+
+class ConcretePublishableFactory(factory.django.DjangoModelFactory):
+
+    class Meta:
+        model = ConcretePublishableModel
